refactor(mespapiers): Split RadioAdapter component
