+++ packages/cozy-mespapiers-lib/src/components/ModelSteps/widgets/RadioAdapter.jsx
@@ -1,15 +1,11 @@
 import React, { Fragment, useState, useEffect } from 'react'
 
-import ListItemText from 'cozy-ui/transpiled/react/ListItemText'
 import Divider from 'cozy-ui/transpiled/react/MuiCozyTheme/Divider'
 import List from 'cozy-ui/transpiled/react/MuiCozyTheme/List'
-import ListItem from 'cozy-ui/transpiled/react/MuiCozyTheme/ListItem'
-import ListItemIcon from 'cozy-ui/transpiled/react/MuiCozyTheme/ListItemIcon'
 import Paper from 'cozy-ui/transpiled/react/Paper'
-import Radio from 'cozy-ui/transpiled/react/Radios'
 
+import RadioAdapterItem from './RadioAdapterItem'
 import { defaultProptypes } from './proptypes'
-import { useScannerI18n } from '../../Hooks/useScannerI18n'
 
 const RadioAdapter = ({
   attrs: { name, options, required },
@@ -18,12 +14,7 @@ const RadioAdapter = ({
   setValidInput,
   setValue
 }) => {
-  const scannerT = useScannerI18n()
   const [currentValue, setCurrentValue] = useState(() => defaultValue || '')
-
-  const isChecked = inputValue => {
-    return inputValue === currentValue
-  }
 
   const handleClick = val => {
     setCurrentValue(val)
@@ -45,18 +36,11 @@ const RadioAdapter = ({
       <List>
         {options.map((option, index) => (
           <Fragment key={index}>
-            <ListItem button onClick={() => handleClick(option)}>
-              <ListItemIcon>
-                <Radio value={option} checked={isChecked(option)} />
-              </ListItemIcon>
-              <ListItemText
-                primary={
-                  <div value={option}>
-                    {scannerT(`attributes.contractType.${option}`)}
-                  </div>
-                }
-              />
-            </ListItem>
+            <RadioAdapterItem
+              option={option}
+              onClick={() => handleClick(option)}
+              value={currentValue}
+            />
             {index !== options.length - 1 && (
               <Divider component="li" variant="inset" />
             )}
