+++ packages/cozy-mespapiers-lib/src/components/ModelSteps/widgets/RadioAdapterItem.jsx
@@ -1,0 +1,40 @@
+import PropTypes from 'prop-types'
+import React from 'react'
+
+import ListItemText from 'cozy-ui/transpiled/react/ListItemText'
+import ListItem from 'cozy-ui/transpiled/react/MuiCozyTheme/ListItem'
+import ListItemIcon from 'cozy-ui/transpiled/react/MuiCozyTheme/ListItemIcon'
+import Radio from 'cozy-ui/transpiled/react/Radios'
+
+import { useScannerI18n } from '../../Hooks/useScannerI18n'
+
+const RadioAdapterItem = ({ onClick, option, value }) => {
+  const scannerT = useScannerI18n()
+
+  const isChecked = inputValue => {
+    return inputValue === value
+  }
+
+  return (
+    <ListItem button onClick={onClick}>
+      <ListItemIcon>
+        <Radio value={option} checked={isChecked(option)} />
+      </ListItemIcon>
+      <ListItemText
+        primary={
+          <div value={option}>
+            {scannerT(`attributes.contractType.${option}`)}
+          </div>
+        }
+      />
+    </ListItem>
+  )
+}
+
+RadioAdapterItem.propTypes = {
+  onClick: PropTypes.func,
+  option: PropTypes.string,
+  value: PropTypes.string
+}
+
+export default RadioAdapterItem
